refactor(frontend): consolidate SubmitButton into shared components
