+++ frontend/feature/users/components/index.ts
@@ -7,5 +7,3 @@ export * from './profile/profile-info';
 export * from './profile/avatar-form';
 export * from './profile/email-form';
 export * from './profile/password-form';
-
-export * from './submit-button';
